+++ pkg/kube/controllers/quarkssecret/quarkssecret_reconciler.go
@@ -2,34 +2,24 @@ package quarkssecret
 
 import (
 	"context"
-	"encoding/base64"
 	"fmt"
-	"strconv"
-	"strings"
 	"time"
 
 	"github.com/pkg/errors"
-	certv1 "k8s.io/api/certificates/v1beta1"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	"code.cloudfoundry.org/quarks-secret/pkg/credsgen"
-	"code.cloudfoundry.org/quarks-secret/pkg/helm/template"
 	qsv1a1 "code.cloudfoundry.org/quarks-secret/pkg/kube/apis/quarkssecret/v1alpha1"
-	"code.cloudfoundry.org/quarks-secret/pkg/kube/util/mutate"
 	"code.cloudfoundry.org/quarks-utils/pkg/config"
 	"code.cloudfoundry.org/quarks-utils/pkg/ctxlog"
 	"code.cloudfoundry.org/quarks-utils/pkg/meltdown"
-	"code.cloudfoundry.org/quarks-utils/pkg/names"
 	"code.cloudfoundry.org/quarks-utils/pkg/pointers"
 )
 
@@ -212,356 +202,6 @@ func (r *ReconcileQuarksSecret) updateStatus(ctx context.Context, qsec *qsv1a1.Q
 	}
 }
 
-func (r *ReconcileQuarksSecret) createPasswordSecret(ctx context.Context, qsec *qsv1a1.QuarksSecret) error {
-	request := credsgen.PasswordGenerationRequest{}
-	password := r.generator.GeneratePassword(qsec.GetName(), request)
-
-	secret := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:        qsec.Spec.SecretName,
-			Namespace:   qsec.GetNamespace(),
-			Labels:      qsec.Spec.SecretLabels,
-			Annotations: qsec.Spec.SecretAnnotations,
-		},
-		StringData: map[string]string{
-			"password": password,
-		},
-	}
-
-	return r.createSecrets(ctx, qsec, secret)
-}
-
-func (r *ReconcileQuarksSecret) createRSASecret(ctx context.Context, qsec *qsv1a1.QuarksSecret) error {
-	key, err := r.generator.GenerateRSAKey(qsec.GetName())
-	if err != nil {
-		return err
-	}
-
-	secret := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:        qsec.Spec.SecretName,
-			Namespace:   qsec.GetNamespace(),
-			Labels:      qsec.Spec.SecretLabels,
-			Annotations: qsec.Spec.SecretAnnotations,
-		},
-		StringData: map[string]string{
-			"private_key": string(key.PrivateKey),
-			"public_key":  string(key.PublicKey),
-		},
-	}
-
-	return r.createSecrets(ctx, qsec, secret)
-}
-
-// renderSecret uses the specified templating engine to draw the secret data
-func (r *ReconcileQuarksSecret) renderSecret(ctx context.Context, qsec *qsv1a1.QuarksSecret) (map[string]string, error) {
-	empty := map[string]string{}
-	if qsec.Spec.TemplateType == "" {
-		return empty, errors.New("templatedConfig needs a templateType to be specified. E.g. helm")
-	}
-
-	// Interpolate the given values, and retrieve the secret contents to fill our config with
-	values := map[string]interface{}{}
-	for k, v := range qsec.Spec.TemplateValues {
-		fields := strings.Split(v.(string), ".")
-		if len(fields) != 2 {
-			return empty, errors.New("failed while reading templatedConfig values. Values must be in the `secretname.field` format. Got: " + v.(string))
-		}
-		secretName := fields[0]
-		field := fields[1]
-
-		existingSecret := &corev1.Secret{}
-		namespacedName := types.NamespacedName{
-			Namespace: qsec.Namespace,
-			Name:      secretName,
-		}
-		err := r.client.Get(ctx, namespacedName, existingSecret)
-		if err != nil {
-			if apierrors.IsNotFound(err) {
-				return empty, newSecNotReadyError("secret not found")
-			}
-			return empty, errors.Wrap(err, "getting secret")
-		}
-		data, ok := existingSecret.Data[field]
-		if !ok {
-			return empty, errors.Errorf("Failed to get secret data key: %s", field)
-		}
-		values[k] = string(data)
-	}
-
-	// Call the specified rendering engine to draw our data
-	switch qsec.Spec.TemplateType {
-	case HelmTemplate:
-		t := template.New()
-		return t.ExecuteMap(qsec.Spec.Template, values), nil
-	default:
-		return map[string]string{}, errors.New("unsupported template type has been specified")
-	}
-}
-
-func (r *ReconcileQuarksSecret) createTemplatedSecret(ctx context.Context, qsec *qsv1a1.QuarksSecret) error {
-	secretData, err := r.renderSecret(ctx, qsec)
-	if err != nil {
-		return err
-	}
-
-	secret := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:        qsec.Spec.SecretName,
-			Namespace:   qsec.GetNamespace(),
-			Labels:      qsec.Spec.SecretLabels,
-			Annotations: qsec.Spec.SecretAnnotations,
-		},
-		StringData: secretData,
-	}
-
-	return r.createSecrets(ctx, qsec, secret)
-}
-
-func (r *ReconcileQuarksSecret) createSSHSecret(ctx context.Context, qsec *qsv1a1.QuarksSecret) error {
-	key, err := r.generator.GenerateSSHKey(qsec.GetName())
-	if err != nil {
-		return err
-	}
-
-	secret := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:        qsec.Spec.SecretName,
-			Namespace:   qsec.GetNamespace(),
-			Labels:      qsec.Spec.SecretLabels,
-			Annotations: qsec.Spec.SecretAnnotations,
-		},
-		StringData: map[string]string{
-			"private_key":            string(key.PrivateKey),
-			"public_key":             string(key.PublicKey),
-			"public_key_fingerprint": key.Fingerprint,
-		},
-	}
-
-	return r.createSecrets(ctx, qsec, secret)
-}
-
-func (r *ReconcileQuarksSecret) createCertificateSecret(ctx context.Context, qsec *qsv1a1.QuarksSecret) error {
-	serviceIPForEKSWorkaround := ""
-
-	for _, serviceRef := range qsec.Spec.Request.CertificateRequest.ServiceRef {
-		service := &corev1.Service{}
-
-		err := r.client.Get(ctx, types.NamespacedName{Namespace: qsec.Namespace, Name: serviceRef.Name}, service)
-
-		if err != nil {
-			return errors.Wrapf(err, "Failed to get service reference '%s' for QuarksSecret '%s'", serviceRef.Name, qsec.GetNamespacedName())
-		}
-
-		if serviceIPForEKSWorkaround == "" {
-			serviceIPForEKSWorkaround = service.Spec.ClusterIP
-		}
-
-		qsec.Spec.Request.CertificateRequest.AlternativeNames = append(append(
-			qsec.Spec.Request.CertificateRequest.AlternativeNames,
-			service.Name,
-			service.Name+"."+service.Namespace,
-			"*."+service.Name,
-			"*."+service.Name+"."+service.Namespace,
-			service.Spec.ClusterIP,
-			service.Spec.LoadBalancerIP,
-			service.Spec.ExternalName,
-		), service.Spec.ExternalIPs...)
-	}
-
-	if len(qsec.Spec.Request.CertificateRequest.SignerType) == 0 {
-		qsec.Spec.Request.CertificateRequest.SignerType = qsv1a1.LocalSigner
-	}
-
-	generationRequest, err := r.generateCertificateGenerationRequest(ctx, qsec.Namespace, qsec.Spec.Request.CertificateRequest)
-	if err != nil {
-		return errors.Wrap(err, "generating certificate generation request")
-	}
-
-	switch qsec.Spec.Request.CertificateRequest.SignerType {
-	case qsv1a1.ClusterSigner:
-		if qsec.Spec.Type == "tls" {
-			return errors.Errorf("can't generate tls Type with cluster SignerType")
-		}
-
-		if qsec.Spec.Request.CertificateRequest.ActivateEKSWorkaroundForSAN {
-			if serviceIPForEKSWorkaround == "" {
-				return errors.Errorf("can't activate EKS workaround for QuarksSecret '%s'; couldn't find a ClusterIP for any service reference", qsec.GetNamespacedName())
-			}
-
-			ctxlog.Infof(ctx, "Activating EKS workaround for QuarksSecret '%s'. Using IP '%s' as a common name. See 'https://github.com/awslabs/amazon-eks-ami/issues/341' for more details.", qsec.GetNamespacedName(), serviceIPForEKSWorkaround)
-
-			generationRequest.CommonName = serviceIPForEKSWorkaround
-		}
-
-		ctxlog.Info(ctx, "Generating certificate signing request and its key")
-		csr, key, err := r.generator.GenerateCertificateSigningRequest(generationRequest)
-		if err != nil {
-			return err
-		}
-
-		// private key Secret which will be merged to certificate Secret later
-		secret := &corev1.Secret{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:        names.CsrPrivateKeySecretName(names.CSRName(qsec.Namespace, qsec.Name)),
-				Namespace:   qsec.GetNamespace(),
-				Labels:      qsec.Spec.SecretLabels,
-				Annotations: qsec.Spec.SecretAnnotations,
-			},
-			StringData: map[string]string{
-				"private_key": string(key),
-				"is_ca":       strconv.FormatBool(qsec.Spec.Request.CertificateRequest.IsCA),
-			},
-		}
-
-		err = r.createSecrets(ctx, qsec, secret)
-		if err != nil {
-			return err
-		}
-
-		return r.createCertificateSigningRequest(ctx, qsec, csr)
-	case qsv1a1.LocalSigner:
-		// Generate certificate
-		cert, err := r.generator.GenerateCertificate(qsec.GetName(), generationRequest)
-		if err != nil {
-			return err
-		}
-
-		objectMeta := metav1.ObjectMeta{
-			Name:        qsec.Spec.SecretName,
-			Namespace:   qsec.GetNamespace(),
-			Labels:      qsec.Spec.SecretLabels,
-			Annotations: qsec.Spec.SecretAnnotations,
-		}
-
-		var secret *corev1.Secret
-		// "tls" QuarksSecret type
-		if qsec.Spec.Type == "tls" {
-			secret = &corev1.Secret{
-				ObjectMeta: objectMeta,
-				StringData: map[string]string{
-					"tls.crt": string(cert.Certificate),
-					"tls.key": string(cert.PrivateKey),
-				},
-				Type: corev1.SecretTypeTLS,
-			}
-		} else {
-			// "certificate" QuarksSecret type
-			secret = &corev1.Secret{
-				ObjectMeta: objectMeta,
-				StringData: map[string]string{
-					"certificate": string(cert.Certificate),
-					"private_key": string(cert.PrivateKey),
-					"is_ca":       strconv.FormatBool(qsec.Spec.Request.CertificateRequest.IsCA),
-				},
-			}
-		}
-
-		if len(generationRequest.CA.Certificate) > 0 {
-			secret.StringData["ca"] = string(generationRequest.CA.Certificate)
-		}
-
-		return r.createSecrets(ctx, qsec, secret)
-	default:
-		return fmt.Errorf("unrecognized signer type: %s", qsec.Spec.Request.CertificateRequest.SignerType)
-	}
-}
-
-func (r *ReconcileQuarksSecret) createBasicAuthSecret(ctx context.Context, qsec *qsv1a1.QuarksSecret) error {
-	username := qsec.Spec.Request.BasicAuthRequest.Username
-	if username == "" {
-		username = r.generator.GeneratePassword(fmt.Sprintf("%s/username", qsec.Name), credsgen.PasswordGenerationRequest{})
-	}
-	password := r.generator.GeneratePassword(fmt.Sprintf("%s/password", qsec.Name), credsgen.PasswordGenerationRequest{})
-
-	secret := &corev1.Secret{
-		Type: corev1.SecretTypeBasicAuth,
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      qsec.Spec.SecretName,
-			Namespace: qsec.GetNamespace(),
-		},
-		StringData: map[string]string{
-			"username": username,
-			"password": password,
-		},
-	}
-
-	return r.createSecrets(ctx, qsec, secret)
-}
-
-func (r *ReconcileQuarksSecret) createDockerConfigJSON(ctx context.Context, qsec *qsv1a1.QuarksSecret) error {
-	// Fetch username and password.
-	username := ""
-	if len(qsec.Spec.Request.ImageCredentialsRequest.Username.Name) > 0 {
-		userSecret := &corev1.Secret{}
-		userNamespacedName := types.NamespacedName{
-			Namespace: qsec.Namespace,
-			Name:      qsec.Spec.Request.ImageCredentialsRequest.Username.Name,
-		}
-		err := r.client.Get(ctx, userNamespacedName, userSecret)
-		if err != nil {
-			if apierrors.IsNotFound(err) {
-				return newSecNotReadyError("username secret not found")
-			}
-			return errors.Wrap(err, "getting username secret")
-		}
-		data, ok := userSecret.Data[qsec.Spec.Request.ImageCredentialsRequest.Username.Key]
-		if !ok {
-			return errors.Errorf("Failed to get username data by key: %s", qsec.Spec.Request.ImageCredentialsRequest.Username.Key)
-		}
-		username = string(data)
-	}
-	if username == "" {
-		username = r.generator.GeneratePassword(fmt.Sprintf("%s/username", qsec.Name), credsgen.PasswordGenerationRequest{})
-	}
-
-	password := ""
-	if len(qsec.Spec.Request.ImageCredentialsRequest.Password.Name) > 0 {
-		passSecret := &corev1.Secret{}
-		passNamespacedName := types.NamespacedName{
-			Namespace: qsec.Namespace,
-			Name:      qsec.Spec.Request.ImageCredentialsRequest.Password.Name,
-		}
-		err := r.client.Get(ctx, passNamespacedName, passSecret)
-		if err != nil {
-			if apierrors.IsNotFound(err) {
-				return newSecNotReadyError("password secret not found")
-			}
-			return errors.Wrap(err, "getting password secret")
-		}
-		data, ok := passSecret.Data[qsec.Spec.Request.ImageCredentialsRequest.Password.Key]
-		if !ok {
-			return errors.Errorf("Failed to get password data key: %s", qsec.Spec.Request.ImageCredentialsRequest.Password.Key)
-		}
-		password = string(data)
-	}
-	if password == "" {
-		password = r.generator.GeneratePassword(fmt.Sprintf("%s/password", qsec.Name), credsgen.PasswordGenerationRequest{})
-	}
-
-	authEncode := base64.StdEncoding.EncodeToString([]byte(fmt.Sprintf("%s:%s", username, password)))
-	dockerConfigJSONData := fmt.Sprintf("{\"auths\":{\"%s\":{\"username\":\"%s\",\"password\":\"%s\",\"email\":\"%s\",\"auth\":\"%s\"}}}",
-		qsec.Spec.Request.ImageCredentialsRequest.Registry,
-		username,
-		password,
-		qsec.Spec.Request.ImageCredentialsRequest.Email,
-		authEncode,
-	)
-
-	secret := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      qsec.Spec.SecretName,
-			Namespace: qsec.GetNamespace(),
-		},
-		Type: corev1.SecretTypeDockerConfigJson,
-		StringData: map[string]string{
-			corev1.DockerConfigJsonKey: dockerConfigJSONData,
-		},
-	}
-
-	return r.createSecrets(ctx, qsec, secret)
-}
-
 // Skip creation when
 // * secret is already generated according to qsecs status field
 // * secret exists, but was not generated (user created secret)
@@ -604,115 +244,6 @@ func (r *ReconcileQuarksSecret) skipCreation(ctx context.Context, qsec *qsv1a1.Q
 	}
 
 	return false, nil
-}
-
-// Skip copy creation when
-// * secret and qsecret (copy type) doesn't exist
-// * (q)secret exists, but it's not marked as generated (label) and a copy (annotation)
-func (r *ReconcileQuarksSecret) skipCopy(ctx context.Context, qsec *qsv1a1.QuarksSecret, sourceQsec *qsv1a1.QuarksSecret) (bool, bool, error) {
-	copyOf := sourceQsec.GetNamespacedName()
-	if qsec.Status.Generated != nil && *qsec.Status.Generated {
-		ctxlog.Debugf(ctx, "Existing secret %s/%s has already been generated",
-			qsec.Namespace,
-			qsec.Spec.SecretName,
-		)
-		return true, false, nil
-	}
-
-	secretName := qsec.Spec.SecretName
-	notFoundQsec := false
-	notFoundSec := false
-
-	// We use an unstructured object so we don't hit the _namespaced_ cache
-	// since our object could live in another namespace
-	existingSecret := &unstructured.Unstructured{}
-	existingSecret.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   "",
-		Kind:    "Secret",
-		Version: "v1",
-	})
-
-	// We want to create the secret if a qsec is present, or update a secret if the secret is already there (with the correct annotation)
-	existingQSec := &qsv1a1.QuarksSecret{}
-	err := r.client.Get(ctx, types.NamespacedName{Name: secretName, Namespace: qsec.GetNamespace()}, existingQSec)
-	if err != nil {
-		if apierrors.IsNotFound(err) {
-			notFoundQsec = true
-		} else {
-			return false, false, errors.Wrapf(err, "could not get qsecret")
-		}
-	}
-
-	err = r.client.Get(ctx, types.NamespacedName{Name: secretName, Namespace: qsec.GetNamespace()}, existingSecret)
-	if err != nil {
-		if apierrors.IsNotFound(err) {
-			notFoundSec = true
-		} else {
-			return false, false, errors.Wrapf(err, "could not get secret")
-		}
-	}
-
-	// If both are absent, we will skip the copy
-	if notFoundSec && notFoundQsec {
-		ctxlog.WithEvent(sourceQsec, "SkipCreation").Infof(ctx, "No Valid QSecret or Secret found")
-		return true, false, nil
-	}
-
-	if notFoundSec {
-		// Validate QSecret
-		ctxlog.WithEvent(sourceQsec, "SkipCreation").Infof(ctx, "Valid QSecret found")
-
-		secretLabels := existingQSec.GetLabels()
-		if secretLabels == nil {
-			secretLabels = map[string]string{}
-		}
-
-		secretAnnotations := existingQSec.GetAnnotations()
-		if secretAnnotations == nil {
-			secretAnnotations = map[string]string{}
-		}
-
-		if existingQSec.Spec.Type != qsv1a1.SecretCopy {
-			ctxlog.WithEvent(sourceQsec, "SkipCreation").Infof(ctx, "Invalid type for QSecret. It must be 'copy' type.")
-			return true, false, nil
-		}
-
-		return !validateCopySecret(ctx, sourceQsec, secretLabels, secretAnnotations, copyOf), true, nil
-	} else if notFoundQsec {
-		// Validate Secret
-		secretLabels := existingSecret.GetLabels()
-		if secretLabels == nil {
-			secretLabels = map[string]string{}
-		}
-
-		secretAnnotations := existingSecret.GetAnnotations()
-		if secretAnnotations == nil {
-			secretAnnotations = map[string]string{}
-		}
-
-		return !validateCopySecret(ctx, sourceQsec, secretLabels, secretAnnotations, copyOf), false, nil
-	}
-
-	return true, false, nil
-}
-
-func validateCopySecret(ctx context.Context, qsec *qsv1a1.QuarksSecret, secretLabels, secretAnnotations map[string]string, copyOf string) bool {
-
-	valid := true
-
-	// check if the secret is marked as generated
-	if secretLabels[qsv1a1.LabelKind] != qsv1a1.GeneratedSecretKind {
-		ctxlog.WithEvent(qsec, "SkipCopyCreation").Infof(ctx, "Secret doesn't have generated label")
-		valid = false
-	}
-
-	// skip if this is a copy, and we're missing a copy-of label
-	if secretAnnotations[qsv1a1.AnnotationCopyOf] != copyOf {
-		ctxlog.WithEvent(qsec, "SkipCopyCreation").Infof(ctx, "Secret doesn't have the corresponding annotation %s vs %s", secretAnnotations[qsv1a1.AnnotationCopyOf], copyOf)
-		valid = false
-	}
-
-	return valid
 }
 
 func (r *ReconcileQuarksSecret) createSecrets(ctx context.Context, qsec *qsv1a1.QuarksSecret, secret *corev1.Secret) error {
@@ -764,200 +295,5 @@ func (r *ReconcileQuarksSecret) createSecrets(ctx context.Context, qsec *qsv1a1.
 		}
 	}
 
-	return nil
-}
-
-// createCopySecret applies common properties(labels and ownerReferences) to the secret and creates it
-func (r *ReconcileQuarksSecret) createCopySecret(ctx context.Context, qsec *qsv1a1.QuarksSecret, secret *corev1.Secret, copyOf string) error {
-	ctxlog.Debugf(ctx, "Creating secret '%s/%s', owned by quarks secret '%s'", secret.Namespace, secret.Name, qsec.GetNamespacedName())
-
-	secretLabels := secret.GetLabels()
-	if secretLabels == nil {
-		secretLabels = map[string]string{}
-	}
-
-	secretAnnotations := secret.GetAnnotations()
-	if secretAnnotations == nil {
-		secretAnnotations = map[string]string{}
-	}
-
-	secretLabels[qsv1a1.LabelKind] = qsv1a1.GeneratedSecretKind
-	secretAnnotations[qsv1a1.AnnotationCopyOf] = copyOf
-
-	secret.SetLabels(secretLabels)
-	secret.SetAnnotations(secretAnnotations)
-	if err := r.setReference(qsec, secret, r.scheme); err != nil {
-		return errors.Wrapf(err, "error setting owner for secret '%s' to QuarksSecret '%s'", secret.GetName(), qsec.GetNamespacedName())
-	}
-
-	op, err := controllerutil.CreateOrUpdate(ctx, r.client, secret, mutate.SecretMutateFn(secret))
-	if err != nil {
-		return errors.Wrapf(err, "could not create or update secret '%s/%s'", secret.Namespace, secret.GetName())
-	}
-
-	if op != "unchanged" {
-		ctxlog.Debugf(ctx, "Secret '%s' has been %s", secret.Name, op)
-	}
-
-	return nil
-}
-
-// createSecret applies common properties(labels and ownerReferences) to the secret and creates it
-func (r *ReconcileQuarksSecret) createSecret(ctx context.Context, qsec *qsv1a1.QuarksSecret, secret *corev1.Secret) error {
-	ctxlog.Debugf(ctx, "Creating secret '%s/%s', owned by quarks secret '%s'", secret.Namespace, secret.Name, qsec.GetNamespacedName())
-
-	secretLabels := secret.GetLabels()
-	if secretLabels == nil {
-		secretLabels = map[string]string{}
-	}
-
-	secretLabels[qsv1a1.LabelKind] = qsv1a1.GeneratedSecretKind
-
-	secret.SetLabels(secretLabels)
-
-	if err := r.setReference(qsec, secret, r.scheme); err != nil {
-		return errors.Wrapf(err, "error setting owner for secret '%s' to QuarksSecret '%s'", secret.GetName(), qsec.GetNamespacedName())
-	}
-
-	op, err := controllerutil.CreateOrUpdate(ctx, r.client, secret, mutate.SecretMutateFn(secret))
-	if err != nil {
-		return errors.Wrapf(err, "could not create or update secret '%s/%s'", secret.Namespace, secret.GetName())
-	}
-
-	if op != "unchanged" {
-		ctxlog.Debugf(ctx, "Secret '%s' has been %s", secret.Name, op)
-	}
-
-	return nil
-}
-
-// updateCopySecret updates a copied destination Secret
-func (r *ReconcileQuarksSecret) updateCopySecret(ctx context.Context, qsec *qsv1a1.QuarksSecret, secret *corev1.Secret) error {
-	// If this is a copy (lives in a different namespace), we only do an update,
-	// since we're not allowed to create, and we don't set a reference, because
-	// cross namespace references are not supported
-	uncachedSecret := &unstructured.Unstructured{}
-	uncachedSecret.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   "",
-		Kind:    "Secret",
-		Version: "v1",
-	})
-	uncachedSecret.SetName(secret.Name)
-	uncachedSecret.SetNamespace(secret.Namespace)
-	uncachedSecret.Object["data"] = secret.Data
-	err := r.client.Update(ctx, uncachedSecret)
-
-	if err != nil {
-		return errors.Wrapf(err, "could not update secret '%s/%s'", secret.Namespace, secret.GetName())
-	}
-
-	return nil
-}
-
-// generateCertificateGenerationRequest generates CertificateGenerationRequest for certificate
-func (r *ReconcileQuarksSecret) generateCertificateGenerationRequest(ctx context.Context, namespace string, certificateRequest qsv1a1.CertificateRequest) (credsgen.CertificateGenerationRequest, error) {
-	var request credsgen.CertificateGenerationRequest
-	switch certificateRequest.SignerType {
-	case qsv1a1.ClusterSigner:
-		// Generate cluster-signed CA certificate
-		request = credsgen.CertificateGenerationRequest{
-			CommonName:       certificateRequest.CommonName,
-			AlternativeNames: certificateRequest.AlternativeNames,
-		}
-	case qsv1a1.LocalSigner:
-		// Generate local-issued CA certificate
-		request = credsgen.CertificateGenerationRequest{
-			IsCA:             certificateRequest.IsCA,
-			CommonName:       certificateRequest.CommonName,
-			AlternativeNames: certificateRequest.AlternativeNames,
-		}
-
-		if len(certificateRequest.CARef.Name) > 0 {
-			// Get CA certificate
-			caSecret := &corev1.Secret{}
-			caNamespacedName := types.NamespacedName{
-				Namespace: namespace,
-				Name:      certificateRequest.CARef.Name,
-			}
-			err := r.client.Get(ctx, caNamespacedName, caSecret)
-			if err != nil {
-				if apierrors.IsNotFound(err) {
-					return request, newCaNotReadyError("CA secret not found")
-				}
-				return request, errors.Wrap(err, "getting CA secret")
-			}
-			ca := caSecret.Data[certificateRequest.CARef.Key]
-
-			// Get CA key
-			if certificateRequest.CAKeyRef.Name != certificateRequest.CARef.Name {
-				caSecret = &corev1.Secret{}
-				caNamespacedName = types.NamespacedName{
-					Namespace: namespace,
-					Name:      certificateRequest.CAKeyRef.Name,
-				}
-				err = r.client.Get(ctx, caNamespacedName, caSecret)
-				if err != nil {
-					if apierrors.IsNotFound(err) {
-						return request, newCaNotReadyError("CA key secret not found")
-					}
-					return request, errors.Wrap(err, "getting CA Key secret")
-				}
-			}
-			key := caSecret.Data[certificateRequest.CAKeyRef.Key]
-			request.CA = credsgen.Certificate{
-				IsCA:        true,
-				PrivateKey:  key,
-				Certificate: ca,
-			}
-		}
-	default:
-		return request, fmt.Errorf("unrecognized signer type: %s", certificateRequest.SignerType)
-	}
-
-	return request, nil
-}
-
-// createCertificateSigningRequest creates CertificateSigningRequest Object
-func (r *ReconcileQuarksSecret) createCertificateSigningRequest(ctx context.Context, qsec *qsv1a1.QuarksSecret, csr []byte) error {
-	csrName := names.CSRName(qsec.Namespace, qsec.Name)
-	ctxlog.Debugf(ctx, "Creating certificatesigningrequest '%s'", csrName)
-
-	annotations := qsec.GetAnnotations()
-	if annotations == nil {
-		annotations = map[string]string{}
-	}
-	annotations[qsv1a1.AnnotationCertSecretName] = qsec.Spec.SecretName
-	annotations[qsv1a1.AnnotationQSecNamespace] = qsec.Namespace
-	annotations[qsv1a1.AnnotationQSecName] = qsec.Name
-	annotations[qsv1a1.AnnotationMonitoredID] = r.config.MonitoredID
-
-	csrObj := &certv1.CertificateSigningRequest{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:        csrName,
-			Labels:      qsec.Labels,
-			Annotations: annotations,
-		},
-		Spec: certv1.CertificateSigningRequestSpec{
-			Request: csr,
-			Usages:  qsec.Spec.Request.CertificateRequest.Usages,
-		},
-	}
-
-	oldCsrObj := &certv1.CertificateSigningRequest{}
-
-	// CSR spec is immutable after the request is created
-	err := r.client.Get(ctx, types.NamespacedName{Name: csrObj.Name}, oldCsrObj)
-	if err != nil {
-		if apierrors.IsNotFound(err) {
-			err = r.client.Create(ctx, csrObj)
-			if err != nil {
-				return errors.Wrapf(err, "could not create certificatesigningrequest '%s'", csrObj.Name)
-			}
-			return nil
-		}
-		return errors.Wrapf(err, "could not get certificatesigningrequest '%s'", csrObj.Name)
-	}
-
-	ctxlog.Infof(ctx, "Ignoring immutable CSR '%s'", csrObj.Name)
 	return nil
 }
